Moved the application back into package + blueprint layout.
More info at http://exploreflask.com/en/latest/organizing.html.

Also, switched from a JSON configuration file to a Python format
as per http://flask.pocoo.org/docs/0.12/config/.
