+++ browser/static/js/ie10-viewport-bug-workaround.js
