+++ browser/static/js/npm.js
