+++ browser/static/libs/jquery.js
