+++ browser/templates/island.html
@@ -7,18 +7,18 @@
     <!-- The above 3 meta tags *must* come first in the head; any other head content must come *after* these tags -->
     <meta name="description" content="A file browser for the Starry Expanse project">
     <meta name="author" content="Chris Mumford">
-    <link rel="icon" href="{{ url_for('static', filename='favicon.ico') }}">
+    <link rel="icon" href="{{ url_for('browsing.static', filename='favicon.ico') }}">
 
     <title>Reference Browser: {{island_name}}</title>
 
     <!-- Bootstrap core CSS -->
-    <link href="{{ url_for('static', filename='css/bootstrap.min.css') }}" rel="stylesheet">
+    <link href="{{ url_for('browsing.static', filename='css/bootstrap.min.css') }}" rel="stylesheet">
 
     <!-- IE10 viewport hack for Surface/desktop Windows 8 bug -->
-    <link href="{{ url_for('static', filename='css/ie10-viewport-bug-workaround.css') }}" rel="stylesheet">
+    <link href="{{ url_for('browsing.static', filename='css/ie10-viewport-bug-workaround.css') }}" rel="stylesheet">
 
     <!-- Custom styles for this template -->
-    <link href="{{ url_for('static', filename='css/main.css') }}" rel="stylesheet">
+    <link href="{{ url_for('browsing.static', filename='css/main.css') }}" rel="stylesheet">
 
     <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
     <!--[if lt IE 9]>
@@ -32,8 +32,8 @@
     <div class="page-header">
       <h1>Reference Browser<br><small>The Riven asset browser</small></h1>
       <div>
-        <a class="btn btn-default" href="{{ url_for('islands') }}">All Islands</a>
-        <a class="btn btn-warning pull-right" href="{{ url_for('logout') }}">Sign Off</a>
+        <a class="btn btn-default" href="{{ url_for('browsing.islands') }}">All Islands</a>
+        <a class="btn btn-warning pull-right" href="{{ url_for('browsing.logout') }}">Sign Off</a>
       </div>
     </div>
 
@@ -43,8 +43,8 @@
       {% if position_count < 0 %}
       <h2>Positions ({{position_count}})</h2>
       {% for position in positions %}
-        <a class="btn btn-default position-btn" href="{{ url_for('island', symbol=island_symbol, position=position.id) }}">
-          <img src="{{ url_for('protected', filename=position.thumbnail) }}"
+        <a class="btn btn-default position-btn" href="{{ url_for('browsing.island', symbol=island_symbol, position=position.id) }}">
+          <img src="{{ url_for('browsing.protected', filename=position.thumbnail) }}"
                width="{{thumbnail_width}}" height="{{thumbnail_height}}"><br>
         </a>
       {% endfor %}
@@ -52,14 +52,14 @@
 
       <h2>Viewpoints ({{viewpoint_count}})</h2>
       {% for viewpoint in viewpoints %}
-         <a class="btn btn-default position-btn" href="{{ url_for('viewpoint', symbol=island_symbol, vpt_name=viewpoint.name) }}">
+         <a class="btn btn-default position-btn" href="{{ url_for('browsing.viewpoint', symbol=island_symbol, vpt_name=viewpoint.name) }}">
            {% if use_unveil %}
-           <img src="{{ url_for('static', filename='images/bg.png') }}"
-               data-src="{{ url_for('protected', filename=viewpoint.thumbnail) }}"
-               data-src-retina="{{ url_for('protected', filename=viewpoint.thumbnail2x) }}"
+           <img src="{{ url_for('browsing.static', filename='images/bg.png') }}"
+               data-src="{{ url_for('browsing.protected', filename=viewpoint.thumbnail) }}"
+               data-src-retina="{{ url_for('browsing.protected', filename=viewpoint.thumbnail2x) }}"
                width="{{thumbnail_width}}" height="{{thumbnail_height}}"><br>
            {% else %}
-             <img src="{{ url_for('protected', filename=viewpoint.thumbnail) }}"
+             <img src="{{ url_for('browsing.protected', filename=viewpoint.thumbnail) }}"
                  width="{{thumbnail_width}}" height="{{thumbnail_height}}"><br>
            {% endif %}
            {{ viewpoint.name }}
@@ -72,7 +72,7 @@
     ================================================== -->
     <!-- Placed at the end of the document so the pages load faster -->
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
-    <script src="{{ url_for('static', filename='js/jquery.unveil.js') }}"></script>
+    <script src="{{ url_for('browsing.static', filename='js/jquery.unveil.js') }}"></script>
     <script type="text/javascript">
       $(document).ready(function() {
         $("img").unveil();
@@ -80,8 +80,8 @@
       });
     </script>
     <script>window.jQuery || document.write('<script src="../../assets/js/vendor/jquery.min.js"><\/script>')</script>
-    <script src="{{ url_for('static', filename='js/bootstrap.min.js') }}"></script>
+    <script src="{{ url_for('browsing.static', filename='js/bootstrap.min.js') }}"></script>
     <!-- IE10 viewport hack for Surface/desktop Windows 8 bug -->
-    <script src="{{ url_for('static', filename='js/ie10-viewport-bug-workaround.js') }}"></script>
+    <script src="{{ url_for('browsing.static', filename='js/ie10-viewport-bug-workaround.js') }}"></script>
   </body>
 </html>
