+++ browser/templates/islands.html
@@ -7,18 +7,18 @@
     <!-- The above 3 meta tags *must* come first in the head; any other head content must come *after* these tags -->
     <meta name="description" content="A file browser for the Starry Expanse project">
     <meta name="author" content="Chris Mumford">
-    <link rel="icon" href="{{ url_for('static', filename='favicon.ico') }}">
+    <link rel="icon" href="{{ url_for('browsing.static', filename='favicon.ico') }}">
 
     <title>Reference Browser</title>
 
     <!-- Bootstrap core CSS -->
-    <link href="{{ url_for('static', filename='css/bootstrap.min.css') }}" rel="stylesheet">
+    <link href="{{ url_for('browsing.static', filename='css/bootstrap.min.css') }}" rel="stylesheet">
 
     <!-- IE10 viewport hack for Surface/desktop Windows 8 bug -->
-    <link href="{{ url_for('static', filename='css/ie10-viewport-bug-workaround.css') }}" rel="stylesheet">
+    <link href="{{ url_for('browsing.static', filename='css/ie10-viewport-bug-workaround.css') }}" rel="stylesheet">
 
     <!-- Custom styles for this template -->
-    <link href="{{ url_for('static', filename='css/main.css') }}" rel="stylesheet">
+    <link href="{{ url_for('browsing.static', filename='css/main.css') }}" rel="stylesheet">
 
     <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
     <!--[if lt IE 9]>
@@ -32,8 +32,8 @@
     <div class="page-header">
       <h1>Reference Browser<br><small>The Riven asset browser</small></h1>
       <div>
-        <a class="btn btn-default disabled" href="{{ url_for('islands') }}">All Islands</a>
-        <a class="btn btn-warning pull-right" href="{{ url_for('logout') }}">Sign Off</a>
+        <a class="btn btn-default disabled" href="{{ url_for('browsing.islands') }}">All Islands</a>
+        <a class="btn btn-warning pull-right" href="{{ url_for('browsing.logout') }}">Sign Off</a>
       </div>
     </div>
 
@@ -43,11 +43,11 @@
         <div class="row">
           {%- for island in column -%}
             <div class="col-sm-6">
-              <a class="btn btn-default navbar-btn btn-wrap" href="{{ url_for('island', symbol=island.symbol) }}">
+              <a class="btn btn-default navbar-btn btn-wrap" href="{{ url_for('browsing.island', symbol=island.symbol) }}">
                 <table>
                   <tr>
                     <td>
-                      <img class="img img-thumbnail" src="{{ url_for('protected', filename=island.icon) }}">
+                      <img class="img img-thumbnail" src="{{ url_for('browsing.protected', filename=island.icon) }}">
                     </td>
                     <td>{{ island.title() }}</td>
                   </tr>
@@ -65,9 +65,9 @@
     <!-- Placed at the end of the document so the pages load faster -->
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
     <script>window.jQuery || document.write('<script src="../../assets/js/vendor/jquery.min.js"><\/script>')</script>
-    <script src="{{ url_for('static', filename='js/bootstrap.min.js') }}"></script>
+    <script src="{{ url_for('browsing.static', filename='js/bootstrap.min.js') }}"></script>
     <!-- IE10 viewport hack for Surface/desktop Windows 8 bug -->
-    <script src="{{ url_for('static', filename='js/ie10-viewport-bug-workaround.js') }}"></script>
+    <script src="{{ url_for('browsing.static', filename='js/ie10-viewport-bug-workaround.js') }}"></script>
   </body>
 </html>
