+++ browser/templates/login.html
@@ -2,7 +2,7 @@
   <head>
     <title>Reference Browser: login page</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="{{ url_for('static', filename='css/bootstrap.min.css') }}" rel="stylesheet">
+    <link href="{{ url_for('browsing.static', filename='css/bootstrap.min.css') }}" rel="stylesheet">
   </head>
   <body>
     <div class="container">
