+++ browser/templates/viewpoints.html
@@ -2,15 +2,15 @@
 <html lang = "en">
   <head>
     <title>All Viewpoints in </title>
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='css/main.css') }}">
+    <link rel="stylesheet" type="text/css" href="{{ url_for('browsing.static', filename='css/main.css') }}">
   </head>
   <body>
       <h3>
-         <a href = "{{ url_for('viewpoints') }}">Riven Viewpoints</a>
+         <a href = "{{ url_for('browsing.viewpoints') }}">Riven Viewpoints</a>
       </h3>
       <hr/>
       <p>
-         <a href = "{{ url_for('islands') }}">Riven Islands</a>
+         <a href = "{{ url_for('browsing.islands') }}">Riven Islands</a>
       </p>
       <table>
          <thead>
