+++ config.py
@@ -1,0 +1,1 @@
+DEBUG=False
